requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.0
aioitertools==0.11.0
aiosignal==1.3.1
aiosmtplib==2.0.2
alembic==1.13.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.1.0
Authlib==1.3.2
bcrypt==4.2.0
beautifulsoup4==4.13.3
billiard==4.2.1
blinker==1.8.2
boto3==1.34.131
botocore==1.34.131
celery==5.4.0
certifi==2024.7.4
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.3
Deprecated==1.2.15
distlib==0.3.8
dnspython==2.6.1
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.111.0
fastapi-cli==0.0.5
fastapi-mail==1.4.1
fastapi-sso==0.17.0
filelock==3.15.4
finnhub-python==2.4.22
forex-python==1.8
frozendict==2.4.6
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
identify==2.6.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.5.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
multitasking==0.0.11
mypy==1.10.0
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==2.1.2
oauthlib==3.2.2
openpyxl==3.1.5
orjson==3.10.6
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
passlib==1.7.4
pdfkit==1.0.0
peewee==3.17.9
platformdirs==4.2.2
playwright==1.51.0
pluggy==1.5.0
pre-commit==3.7.1
prompt_toolkit==3.0.50
psycopg==3.1.19
psycopg-binary==3.1.19
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.7.4
pydantic-settings==2.3.3
pydantic_core==2.18.4
pyee==12.1.1
PyGithub==2.5.0
Pygments==2.18.0
PyJWT==2.8.0
PyNaCl==1.5.0
PySocks==1.7.1
pytest==7.4.3
pytest-asyncio==0.23.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
ruff==0.4.9
s3transfer==0.10.2
selenium==4.29.0
shellingham==1.5.4
simplejson==3.20.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.31
starlette==0.37.2
tenacity==9.1.2
tomli==2.0.1
trio==0.25.1
trio-websocket==0.12.2
typer==0.12.3
types-passlib==1.7.7.20240106
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
undetected-chromedriver==3.5.5
urllib3==2.2.2
uvicorn==0.24.0.post1
vine==5.1.0
virtualenv==20.26.3
watchfiles==0.22.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
yfinance==0.2.54
